% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{sparkreadr}
\alias{sparkreadr}
\alias{sparkreadr-package}
\title{Read in different data files into Spark}
\description{
'sparkreadr' uses several Spark packages to load datasets in parallel in Spark.
}

