% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{sparkreadr}
\alias{sparkreadr}
\alias{sparkreadr-package}
\title{Read in flat data files into Spark}
\description{
'sparkreadr' is an extension for sparklyr to read flat files into Spark DataFrames. It uses the Spark package spark-csv.
}

